# static/swagger.yaml
swagger: "2.0"
info:
  title: httpbin.org
  description: A simple HTTP Request & Response Service.
  version: 0.0.1
host: localhost:8080
schemes:
  http
basePath: /
produces:
  application/json
paths:
...
  /redirect-to:
    delete:
      produces:
        - text/html
      responses:
        '404':
          description: Redirect to donn't support delete, because it does not has parameter url.
      summary: 404 Not Found
      tags:
        - Redirects
    get:
      parameters:
        - in: query
          name: url
          required: true
          type: string
        - in: query
          name: status_code
          type: int
      produces:
        - text/html
      responses:
        '302':
          description: A redirection.
      summary: 302/3XX Redirects to the given URL.
      tags:
        - Redirects
    post:
      parameters:
        - in: formData
          name: url
          required: true
          type: string
        - in: formData
          name: status_code
          required: false
          type: int
      produces:
        - text/html
      responses:
        '302':
          description: A redirection.
      summary: 302/3XX Redirects to the given URL.
      tags:
        - Redirects
    patch:
      parameters:
        - in: formData
          name: url
          required: true
          type: string
        - in: formData
          name: status_code
          required: false
          type: int
      produces:
        - text/html
      responses:
        '302':
          description: A redirection.
      summary: 302/3XX Redirects to the given URL.
      tags:
        - Redirects
    put:
      parameters:
        - in: formData
          name: url
          required: true
          type: string
        - in: formData
          name: status_code
          required: false
          type: int
      produces:
        - text/html
      responses:
        '302':
          description: A redirection.
      summary: 302/3XX Redirects to the given URL.
      tags:
        - Redirects
...